New Renderer3D works now well enough.

+++ tests/src/SkinnedMeshTestScene.cpp
@@ -187,24 +187,8 @@ void SkinnedMeshTestScene::init()
     Model* pBoxModel = pAssetManager->loadModel("assets/TestCube.glb");
 
 
-    int area = 11;
+    int area = 2;
     float spacing = 3.25f;
-    /*
-    for (int x = 0; x < area; ++x)
-    {
-        for (int z = 0; z < area; ++z)
-        {
-            create_animated_entity(
-                this,
-                pAnimatedMesh,
-                bindPoses[0],
-                pAnimationAsset,
-                pMaterial,
-                { x * spacing, 0, -z * spacing }
-            );
-        }
-    }
-    */
 
     for (int x = 0; x < area; ++x)
     {
@@ -223,7 +207,6 @@ void SkinnedMeshTestScene::init()
                 _jointEntities
             );
 
-            /*
             const float cubeScale = 0.3f;
             entityID_t boxEntity = createEntity();
             create_transform(
@@ -265,7 +248,6 @@ void SkinnedMeshTestScene::init()
                 bindPoses[0],
                 "hand1"
             );
-            */
         }
     }
 
@@ -274,9 +256,6 @@ void SkinnedMeshTestScene::init()
         ComponentType::COMPONENT_TYPE_DIRECTIONAL_LIGHT
     );
     pDirLight->direction = { 0.75f, -1.5f, 1.0f };
-
-
-    Debug::log("___TEST___SKINNED TEST INIT! Created " + std::to_string(getEntities().size()) + " entities.");
 }
 
 static std::string get_joint_name(const Pose& bindPose, size_t index)
@@ -302,12 +281,9 @@ void SkinnedMeshTestScene::update()
 
     InputManager& inputManager = Application::get_instance()->getInputManager();
     if (inputManager.isKeyDown(KeyName::KEY_0) && !s_tabDown)
-    {
         Application::get_instance()->getSceneManager().assignNextScene(new MaterialTestScene);
-    }
 
     // Test removing entities, especially from the middle of the hierarchy
-    /*
     if (inputManager.isKeyDown(KeyName::KEY_TAB) && !s_tabDown)
     {
         s_tabDown = true;
@@ -329,9 +305,11 @@ void SkinnedMeshTestScene::update()
         Debug::log(
             "___TEST___deleting joint: " + get_joint_name(_bindPose, _selectedJointIndex)
         );
-        destroyEntity(_selectedJointEntity);
+        if (entityExists(_selectedJointEntity))
+            destroyEntity(_selectedJointEntity);
+        else
+            Debug::log("___TEST___Entity has already been destroyed!");
     }
     if (!inputManager.isKeyDown(KeyName::KEY_BACKSPACE) && s_backspaceDown)
         s_backspaceDown = false;
-    */
 }
